CBLmodel.py: Look up chunks by key
Many small changes to enable chunks to be looked up by their key value
in the chunkatory rather than by iterating over every chunk, looking for
a match.

Additionally, keep the max chunk length up to date as we add new chunks
to the collection of all chunks.

The goal of all of this is to speed up utterance reconstruction.